Cleaned up searchspace code, need to replace previoius searchspace definition function.
